Add HTML, CSS and multiple JS files
I created the entire thing on a different file
because using the folder where I added git + webpack,
it was super laggy.

On the external file, I used modules to create the
static pages.

After that I copy-pasta everything to the original
folder and imported images to each module.

Overall the site looks pretty nice. Just laggy because
of the heavy background image.